requirements.txt
google_adk
line-bot-sdk
fastapi
uvicorn
google-api-python-client
sqlalchemy
psycopg2-binary